Fix warnings.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@9250 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/ap/qnx/ap_QNXDialog_Download_File.cpp
@@ -63,7 +63,7 @@ double percent;
 	if(getProgress() == -1)
 	 	return;
 percent = ((double)(getProgress()))/((double)(getFileSize()))*100;
-value = percent;
+value = (long)percent;
 PtSetResource(m_progressBar,Pt_ARG_GAUGE_VALUE,value,0);
 }
